# src/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "API Geekos Heroes of the Storm",
    "description": "API para la administracion de datos de Heroes of the Storm de la comunidad Geekos",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "email": "lanyev@gmail.com"
    },
    "version": "1.0.0"
  },
  "host": "localhost:9000",
  "basePath": "/",
  "schemes": ["http"],
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "paths": {
    "/api/v1/users": {
      "get": {
        "description": "Get a list of all registered users in the Server",
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "404": {
            "description": "User not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/v1/heroes": {
      "get": {
        "description": "Get a list of all heroes in the Heroes of the Storm game",
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "404": {
            "description": "Heroes not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/v1/tank": {
      "get": {
        "description": "Get all entries from the tank table",
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "404": {
            "description": "Entries not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/v1/damage": {
      "get": {
        "description": "Get all entries from the damage table",
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "404": {
            "description": "Entries not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/v1/health": {
      "get": {
        "description": "Get all entries from the health table",
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "404": {
            "description": "Entries not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    }
  }
}